Added gwt-log v3.0.2 library
Pretty cool logging for GWT applications.  Has a bunch of options of
where to show the logs, which are configured in Volinfoman.gwt.xml.
Logs can be turned off for production, and logging statements are
compiled out for optimization.

Note: WindowLogger doesn't work in Firefox if you have open new
windows in tabs set.

Remote logging is set up, so logs are reported back to GAE when the
app is deployed, and you can see them in the logs section of the
dashboard.

Firebug logging doesn't seem to work that well, as when I reload the
page, Firebug closes and reopens, and some of the log entries don't
show up.

- Added gwt-log and its javadocs to /lib
- Added gwt-log javadocs to Eclipse project for gwt-log library
- Added gwt-log to war/WEB-INF/lib
- Added gwt-log configuration to
	src/com/lisedex/volinfoman/Volinfoman.gwt.xml
- Added gwt-log example code to
  	/src/com/lisedex/volinfoman/client/Volinfoman.java
- Added gwt-log log level config to war/Volinfoman.html
- Set application name for App Engine deployment
- Added gwt-log server logging configuration to war/WEB-INF/web.xml
- Left code in state with logging enabled, DEBUG level.  Logging enabled
	in Window, GWT, and System.

Tested local and GAE